+++ spec/provider/BingProvider.spec.ts
@@ -1,7 +1,7 @@
 import UniversalGeocoder from "UniversalGeocoder";
-import { BingProvider } from "provider";
+import { BingGeocoded, BingProvider } from "provider";
 import ExternalLoader from "ExternalLoader";
-import Geocoded from "Geocoded";
+import AdminLevel from "AdminLevel";
 import setupPolly, { cleanRecording } from "../setupPolly";
 
 describe("Bing Geocoder Provider", () => {
@@ -50,19 +50,19 @@ describe("Bing Geocoder Provider", () => {
 
     provider?.geocode(
       "1600 Pennsylvania Ave, Washington, DC",
-      (results: Geocoded[]) => {
+      (results: BingGeocoded[]) => {
         const geocoded = results[0];
 
         expect(geocoded).toBeDefined();
         expect(geocoded.getCoordinates()).toEqual({
-          latitude: 38.897668,
-          longitude: -77.036556,
+          latitude: 38.897639,
+          longitude: -77.036475,
         });
         expect(geocoded.getBounds()).toEqual({
-          latitudeSW: 38.89380528242933,
-          longitudeSW: -77.04317326462667,
-          latitudeNE: 38.90153071757068,
-          longitudeNE: -77.02993873537334,
+          latitudeSW: 38.89377628242932,
+          longitudeSW: -77.04309226192471,
+          latitudeNE: 38.901501717570675,
+          longitudeNE: -77.02985773807528,
         });
         expect(geocoded.getFormattedAddress()).toEqual(
           "1600 Pennsylvania Ave NW, Washington, DC 20006"
@@ -73,9 +73,22 @@ describe("Bing Geocoder Provider", () => {
         expect(geocoded.getLocality()).toEqual("Washington");
         expect(geocoded.getPostalCode()).toEqual("20006");
         expect(geocoded.getRegion()).toEqual("DC");
-        expect(geocoded.getAdminLevels()).toEqual([]);
+        expect(geocoded.getAdminLevels()).toEqual([
+          AdminLevel.create({
+            level: 1,
+            name: "DC",
+          }),
+          AdminLevel.create({
+            level: 2,
+            name: "City of Washington",
+          }),
+        ]);
         expect(geocoded.getCountry()).toEqual("United States");
-        expect(geocoded.getCountryCode()).toEqual(undefined);
+        expect(geocoded.getCountryCode()).toEqual("US");
+        expect(geocoded.getAttribution()).toEqual(
+          "Copyright © 2020 Microsoft and its suppliers. All rights reserved. This API cannot be accessed and the content and any results may not be used, reproduced or transmitted in any manner without express written permission from Microsoft Corporation."
+        );
+        expect(geocoded.getPrecision()).toEqual("Medium");
 
         done();
       }
@@ -89,7 +102,7 @@ describe("Bing Geocoder Provider", () => {
       apiKey: "api_key",
     });
 
-    provider?.geodecode(48.8631507, 2.388911, (results: Geocoded[]) => {
+    provider?.geodecode(48.8631507, 2.388911, (results: BingGeocoded[]) => {
       const geocoded = results[0];
 
       expect(geocoded).toBeDefined();
@@ -112,9 +125,22 @@ describe("Bing Geocoder Provider", () => {
       expect(geocoded.getLocality()).toEqual("Paris");
       expect(geocoded.getPostalCode()).toEqual("75020");
       expect(geocoded.getRegion()).toEqual("Île-de-France");
-      expect(geocoded.getAdminLevels()).toEqual([]);
+      expect(geocoded.getAdminLevels()).toEqual([
+        AdminLevel.create({
+          level: 1,
+          name: "Île-de-France",
+        }),
+        AdminLevel.create({
+          level: 2,
+          name: "Paris",
+        }),
+      ]);
       expect(geocoded.getCountry()).toEqual("France");
-      expect(geocoded.getCountryCode()).toEqual(undefined);
+      expect(geocoded.getCountryCode()).toEqual("FR");
+      expect(geocoded.getAttribution()).toEqual(
+        "Copyright © 2020 Microsoft and its suppliers. All rights reserved. This API cannot be accessed and the content and any results may not be used, reproduced or transmitted in any manner without express written permission from Microsoft Corporation."
+      );
+      expect(geocoded.getPrecision()).toEqual("High");
 
       done();
     });
@@ -134,7 +160,7 @@ describe("Bing Geocoder Provider", () => {
       },
       (error) => {
         expect(error.message).toEqual(
-          "Received HTTP status code 401 when attempting geocoding request."
+          "Access was denied. You may have entered your credentials incorrectly, or you might not have access to the requested resource or operation."
         );
         done();
       }