+++ src/GeocoderProviderFactory.ts
@@ -1,5 +1,6 @@
 import {
   BingProvider,
+  BingProviderOptionsInterface,
   ChainProvider,
   ChainProviderOptionsInterface,
   GeoPluginProvider,
@@ -16,6 +17,7 @@ import {
   YandexProvider,
   YandexProviderOptionsInterface,
   ProviderOptionsInterface,
+  defaultBingProviderOptions,
   defaultChainProviderOptions,
   defaultMapboxProviderOptions,
   defaultMapQuestProviderOptions,
@@ -29,12 +31,14 @@ import ExternalLoader from "ExternalLoader";
 interface ProviderOptionInterface {
   provider:
     | "bing"
+    | "bingmaps"
     | "chain"
     | "geoplugin"
     | "google"
     | "googlemaps"
     | "mapbox"
     | "mapquest"
+    | "microsoft"
     | "nominatim"
     | "opencage"
     | "openstreetmap"
@@ -49,6 +53,12 @@ interface ChainGeocoderProviderFactoryOptions
   extends ProviderOptionInterface,
     ChainProviderOptionsInterface {
   provider: "chain";
+}
+
+interface BingGeocoderProviderFactoryOptions
+  extends ProviderOptionInterface,
+    BingProviderOptionsInterface {
+  provider: "bing" | "bingmaps" | "microsoft";
 }
 
 interface GeoPluginGeocoderProviderFactoryOptions
@@ -95,6 +105,7 @@ interface YandexGeocoderProviderFactoryOptions
 export type GeocoderProviderFactoryOptions =
   | ProviderFactoryOptions
   | ChainGeocoderProviderFactoryOptions
+  | BingGeocoderProviderFactoryOptions
   | GeoPluginGeocoderProviderFactoryOptions
   | GoogleMapsGeocoderProviderFactoryOptions
   | MapboxGeocoderProviderFactoryOptions
@@ -118,6 +129,8 @@ export type GeocoderProviderByOptionsType<
   O
 > = O extends ChainGeocoderProviderFactoryOptions
   ? ChainProvider
+  : O extends BingGeocoderProviderFactoryOptions
+  ? BingProvider
   : O extends GeoPluginGeocoderProviderFactoryOptions
   ? GeoPluginProvider
   : O extends GoogleMapsGeocoderProviderFactoryOptions
@@ -158,8 +171,13 @@ export default class ProviderFactory {
     const { provider, ...providerOptions } = createProviderOptions;
     switch (provider) {
       case "bing":
+      case "bingmaps":
+      case "microsoft":
         return <GeocoderProviderByOptionsType<O>>(
-          new BingProvider(externalLoader, providerOptions)
+          new BingProvider(externalLoader, {
+            ...defaultBingProviderOptions,
+            ...providerOptions,
+          })
         );
       case "chain":
         return <GeocoderProviderByOptionsType<O>>new ChainProvider({