+++ src/provider/opencage/OpenCageGeocoded.ts
@@ -3,6 +3,7 @@ import Geocoded, { GeocodedObject } from "Geocoded";
 export interface OpenCageGeocodedObject extends GeocodedObject {
   readonly callingCode?: number;
   readonly flag?: string;
+  readonly precision?: number;
   readonly mgrs?: string;
   readonly maidenhead?: string;
   readonly geohash?: string;
@@ -13,6 +14,8 @@ export default class OpenCageGeocoded extends Geocoded {
   private readonly callingCode?: number;
 
   private readonly flag?: string;
+
+  private readonly precision?: number;
 
   private readonly mgrs?: string;
 
@@ -25,6 +28,7 @@ export default class OpenCageGeocoded extends Geocoded {
   protected constructor({
     callingCode,
     flag,
+    precision,
     mgrs,
     maidenhead,
     geohash,
@@ -34,6 +38,7 @@ export default class OpenCageGeocoded extends Geocoded {
     super(geocodedObject);
     this.callingCode = callingCode;
     this.flag = flag;
+    this.precision = precision;
     this.mgrs = mgrs;
     this.maidenhead = maidenhead;
     this.geohash = geohash;
@@ -49,6 +54,7 @@ export default class OpenCageGeocoded extends Geocoded {
       ...super.toObject(),
       callingCode: this.callingCode,
       flag: this.flag,
+      precision: this.precision,
       mgrs: this.mgrs,
       maidenhead: this.maidenhead,
       geohash: this.geohash,
@@ -76,6 +82,17 @@ export default class OpenCageGeocoded extends Geocoded {
 
   public getFlag(): undefined | string {
     return this.flag;
+  }
+
+  public withPrecision(precision: number): OpenCageGeocoded {
+    return new OpenCageGeocoded({
+      ...this.toObject(),
+      precision,
+    });
+  }
+
+  public getPrecision(): undefined | number {
+    return this.precision;
   }
 
   public withMgrs(mgrs: string): OpenCageGeocoded {